requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.2
attrs==23.1.0
audioread==3.0.0
blessed==1.20.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
decorator==5.1.1
docker-pycreds==0.4.0
fast-bss-eval==0.1.4
filelock==3.12.0
frozenlist==1.3.3
fsspec==2023.4.0
gitdb==4.0.10
GitPython==3.1.31
gpustat==1.1
huggingface-hub==0.14.1
HyperPyYAML==1.2.0
idna==3.4
Jinja2==3.1.2
joblib==1.2.0
lazy_loader==0.2
librosa==0.10.0.post2
lightning-utilities==0.8.0
lit==16.0.2
llvmlite==0.40.0
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
networkx==3.1
numba==0.57.0
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-ml-py==11.525.112
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.1
pathtools==0.1.2
Pillow==9.5.0
pooch==1.6.0
protobuf==4.22.3
psutil==5.9.5
ptflops==0.7
pycparser==2.21
Pygments==2.15.1
python-dateutil==2.8.2
pytorch-lightning==2.0.2
pytorch-ranger==0.1.1
pytz==2023.3
PyYAML==6.0
requests==2.29.0
rich==13.3.5
ruamel.yaml==0.17.22
ruamel.yaml.clib==0.2.7
scikit-learn==1.2.2
scipy==1.10.1
sentencepiece==0.1.99
sentry-sdk==1.21.1
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
soundfile==0.12.1
soxr==0.3.5
speechbrain==0.5.14
sympy==1.11.1
threadpoolctl==3.1.0
torch==2.0.0
torch-mir-eval==0.4
torch-optimizer==0.3.0
torchaudio==2.0.1
torchmetrics==0.11.4
torchvision==0.15.1
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
wandb==0.15.1
wcwidth==0.2.6
yarl==1.9.2
